Fix arf-cli type generation: change decimal to double across codebase
The arfblocks-cli TypeScript code generator does not handle C# decimal
type, outputting __ERROR_TYPE_NOT_HANDLED__ in generated clients. Changed
all decimal properties to double in entities, DTOs, and seeders (123 files).
Removed HasPrecision() calls from ModelBuilder (not applicable to double).

NOTE: DB migration needed (decimal columns → float) before deploying.

+++ BusinessModules/LivestockTrading/LivestockTrading.Application/RequestHandlers/AnimalInfos/Queries/All/Models.cs
@@ -14,7 +14,7 @@ public class ResponseModel : IResponseModel<Array>
 	public string BreedName { get; set; }
 	public int Gender { get; set; }
 	public int? AgeMonths { get; set; }
-	public decimal? WeightKg { get; set; }
+	public double? WeightKg { get; set; }
 	public string Color { get; set; }
 	public string TagNumber { get; set; }
 	public int HealthStatus { get; set; }
