+++ BusinessModules/LivestockTrading/LivestockTrading.Application/RequestHandlers/Farms/Commands/Update/Models.cs
@@ -9,8 +9,8 @@ public class RequestModel : IRequestModel
 	public Guid SellerId { get; set; }
 	public Guid LocationId { get; set; }
 	public int Type { get; set; }
-	public decimal? TotalAreaHectares { get; set; }
-	public decimal? CultivatedAreaHectares { get; set; }
+	public double? TotalAreaHectares { get; set; }
+	public double? CultivatedAreaHectares { get; set; }
 	public string Certifications { get; set; }
 	public bool IsOrganic { get; set; }
 	public string ImageUrls { get; set; }
@@ -28,8 +28,8 @@ public class ResponseModel : IResponseModel
 	public Guid SellerId { get; set; }
 	public Guid LocationId { get; set; }
 	public int Type { get; set; }
-	public decimal? TotalAreaHectares { get; set; }
-	public decimal? CultivatedAreaHectares { get; set; }
+	public double? TotalAreaHectares { get; set; }
+	public double? CultivatedAreaHectares { get; set; }
 	public string Certifications { get; set; }
 	public bool IsOrganic { get; set; }
 	public string ImageUrls { get; set; }
