+++ BusinessModules/LivestockTrading/LivestockTrading.Application/RequestHandlers/ProductPrices/Queries/All/Models.cs
@@ -12,8 +12,8 @@ public class ResponseModel : IResponseModel<Array>
 	public Guid Id { get; set; }
 	public Guid ProductId { get; set; }
 	public string CurrencyCode { get; set; }
-	public decimal Price { get; set; }
-	public decimal? DiscountedPrice { get; set; }
+	public double Price { get; set; }
+	public double? DiscountedPrice { get; set; }
 	public string CountryCodes { get; set; }
 	public bool IsActive { get; set; }
 	public DateTime? ValidFrom { get; set; }
